Function to create the content in RStudio Connect

Function to create the content in RStudio Connect

create_rsc_content(
  name,
  description,
  title,
  access_type = "acl",
  connection_timeout = NULL,
  idle_timeout = NULL,
  init_timeout = NULL,
  load_factor = NULL,
  max_conns_per_process = NULL,
  max_processes = NULL,
  min_processes = NULL,
  read_timeout = NULL
)

Arguments

name
Name of the application to be used as an identifier

description
A rich description of the content

title
The title of the content

access_type
Access controls for the content. Allowed options are: "acl" (default), "logged_in", "all"

connection_timeout
Maximum number of seconds allowed without data sent or received across a client connection

idle_timeout
The maximum number of seconds a worker process for an interactive application to remain alive after it goes idle

init_timeout
The maximum number of seconds allowed for an interactive application to start

load_factor
Controls how aggressively new processes are spawned

max_conns_per_process
Specifies the maximum number of client connections allowed to an individual process.

max_processes
Specifies the total number of concurrent processes allowed for a single interactive application

min_processes
Specifies the minimum number of concurrent processes allowed for a single interactive application

read_timeout
Maximum number of seconds allowed without data received from a client connection

Value

content_guid

Details

You can take a look at the options here: https://docs.rstudio.com/connect/1.8.6/api/#post-/v1/content